Move extension refresh swap function to libs so other clients can use it
