+++ apps/browser/src/popup/extension-refresh-route-utils.ts
@@ -1,31 +1,8 @@
-import { inject, Type } from "@angular/core";
-import { Route, Router, Routes, UrlTree } from "@angular/router";
+import { inject } from "@angular/core";
+import { Router, UrlTree } from "@angular/router";
 
-import { componentRouteSwap } from "@bitwarden/angular/utils/component-route-swap";
 import { FeatureFlag } from "@bitwarden/common/enums/feature-flag.enum";
 import { ConfigService } from "@bitwarden/common/platform/abstractions/config/config.service";
-
-/**
- * Helper function to swap between two components based on the ExtensionRefresh feature flag.
- * @param defaultComponent - The current non-refreshed component to render.
- * @param refreshedComponent - The new refreshed component to render.
- * @param options - The shared route options to apply to both components.
- */
-export function extensionRefreshSwap(
-  defaultComponent: Type<any>,
-  refreshedComponent: Type<any>,
-  options: Route,
-): Routes {
-  return componentRouteSwap(
-    defaultComponent,
-    refreshedComponent,
-    async () => {
-      const configService = inject(ConfigService);
-      return configService.getFeatureFlag(FeatureFlag.ExtensionRefresh);
-    },
-    options,
-  );
-}
 
 /**
  * Helper function to redirect to a new URL based on the ExtensionRefresh feature flag.
